Refactor. Modify redeem script size validation method to make it general for any script
