+++ rskj-core/src/main/java/co/rsk/peg/bitcoin/Standardness.java
@@ -1,4 +1,4 @@
-package co.rsk.rules;
+package co.rsk.peg.bitcoin;
 
 public class Standardness {
     public static final int MAX_SCRIPT_ELEMENT_SIZE = 520;
