+++ rskj-core/src/test/java/co/rsk/peg/LegacyErpFederationTest.java
@@ -19,7 +19,7 @@ import co.rsk.config.BridgeMainNetConstants;
 import co.rsk.config.BridgeTestNetConstants;
 import co.rsk.peg.bitcoin.BitcoinTestUtils;
 import co.rsk.peg.resources.TestConstants;
-import co.rsk.rules.Standardness;
+import co.rsk.peg.utils.FederationUtils;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -361,7 +361,7 @@ class LegacyErpFederationTest {
         RawGeneratedRedeemScript[] generatedScripts = new ObjectMapper().readValue(rawRedeemScripts, RawGeneratedRedeemScript[].class);
         for (RawGeneratedRedeemScript generatedScript : generatedScripts) {
             // Skip test cases where the redeem script exceeds the maximum size
-            if (generatedScript.script.getProgram().length <= Standardness.MAX_SCRIPT_ELEMENT_SIZE) {
+            if (FederationUtils.isScriptSizeValid(generatedScript.script)) {
                 Federation erpFederation = new LegacyErpFederation(
                     FederationTestUtils.getFederationMembersWithBtcKeys(generatedScript.mainFed),
                     ZonedDateTime.parse("2017-06-10T02:30:00Z").toInstant(),
